backend/requirements.txt
asgiref==3.7.2
cffi==1.16.0
coverage==7.4.4
cryptography==42.0.5
Django==4.2.11
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
Faker==22.5.1
phonenumbers==8.13.33
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pydot==2.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
celery==5.4.0
redis==5.0.4
mock==5.1.0